Return AttributeLock on lock request

+++ projects/views.py
@@ -821,9 +821,9 @@ class AttributeViewSet(viewsets.ReadOnlyModelViewSet):
 
     @extend_schema(
         responses={
-            200: OpenApiTypes.STR,
+            200: AttributeLockSerializer,
             400: OpenApiTypes.STR,
-            403: OpenApiTypes.STR
+            500: OpenApiTypes.STR
         },
     )
     @action(
@@ -841,26 +841,29 @@ class AttributeViewSet(viewsets.ReadOnlyModelViewSet):
             return HttpResponse(status=status.HTTP_400_BAD_REQUEST)
 
         attribute_lock = AttributeLock.objects.filter(project=project, attribute=attribute).first()
-        if attribute_lock:
-            if (datetime.now(timezone.utc) - attribute_lock.timestamp).total_seconds() >= 900:  # 15 minutes
-                attribute_lock.delete()
-            else:
-                return HttpResponse(
-                    status=status.HTTP_403_FORBIDDEN if request.user != attribute_lock.user else status.HTTP_200_OK
-                )
 
-        attribute_lock = AttributeLock.objects.create(
-            project=project,
-            attribute=attribute,
-            user=request.user
-        )
-        return HttpResponse(status=status.HTTP_200_OK)
+        if attribute_lock and (datetime.now(timezone.utc) - attribute_lock.timestamp).total_seconds() >= 900:  # 15 minutes
+            attribute_lock.delete()
+            attribute_lock = None
+
+        if not attribute_lock:
+            attribute_lock = AttributeLock.objects.create(
+                project=project,
+                attribute=attribute,
+                user=request.user
+            )
+
+        if attribute_lock:
+            return Response({
+                "attribute_lock": AttributeLockSerializer(attribute_lock, context={'request': request}).data
+            }, status=status.HTTP_200_OK)
+
+        return HttpResponse(status=status.HTTP_500_INTERNAL_SERVER_ERROR)
 
     @extend_schema(
         responses={
             200: OpenApiTypes.STR,
-            400: OpenApiTypes.STR,
-            403: OpenApiTypes.STR
+            404: OpenApiTypes.STR
         },
     )
     @action(
@@ -880,7 +883,7 @@ class AttributeViewSet(viewsets.ReadOnlyModelViewSet):
             attribute_lock.delete()
             return HttpResponse(status=status.HTTP_200_OK)
         except AttributeLock.DoesNotExist:
-            return HttpResponse(status=status.HTTP_400_BAD_REQUEST)
+            return HttpResponse(status=status.HTTP_404_NOT_FOUND)
 
 
 @extend_schema(
